+++ code/mathematical-algorithms/pythagorean_triplet/pythagorean_triplet.cpp
