+++ code/mathematical-algorithms/replace_0_with_5/0_to_5_efficent.cpp
