+++ code/mathematical-algorithms/replace_0_with_5/replace_0_with_5.py
