+++ code/mathematical-algorithms/russian_peasant_multiplication/russian_peasant_multiplication.cpp
