+++ code/mathematical-algorithms/shuffle_array/shuffle_array.rb
