+++ code/mathematical-algorithms/sieve_of_atkin/sieve_of_atkin.cpp
