+++ code/mathematical-algorithms/sieve_of_eratosthenes/sieve_of_eratosthenes.hs
