+++ code/mathematical-algorithms/simpsons_rule/simpsons_rule.py
