+++ code/mathematical-algorithms/smallest_digit_in_number/Smallest_digit_in_number.hs
