+++ code/mathematical-algorithms/smallest_digit_in_number/smallest_digit_in_number.c
