+++ code/mathematical-algorithms/sum_of_digits/sum_of_digits.c
