+++ code/mathematical-algorithms/tower_of_hanoi/tower_of_hanoi.scala
