+++ code/mathematical-algorithms/tower_of_hanoi/tower_of_hanoi_binary_solution.c
